cache for contract info
