+++ example/context/config.toml
@@ -6,5 +6,5 @@ ContractManagerAddress = "testnet.confluxscan.io"
 ERC20Address = "CFXTEST:TYPE.CONTRACT:ACDDTMBZ2ZTGAD5KY4P62K0J51RDMZV6Z6HXF0FNSR"
 ERC777Address = "CFXTEST:TYPE.CONTRACT:ACBG83W3HYZC90U5SH6MNRR85X70KVS6SPZN3ZHX84"
 NormalTransactions = ["0xc765d5039ec398c12ba78a8fa6f59f2a0b93324ec0e7db4fa5f9f8aec9aaff87"]
-ERC20Transactions = ["0x587cdb85b1f6b65d2e198b2c908da3b12a2fd3c286ab89628cd168dc8503e89d", "0x9cc78cf06c42abfa6c246fc0dd62f1cc0a2511eb5e53c4342876547bc4c1f5bc", "0x018fb70250123ba3eb5b492ff5fdd76a7360b539031953bd2fcc91ee71c15604", "0x2a9c8417dca16f7026fa465f7bb2cd6d7e845ea0dd773cfd961647381da0c5d8", "0x3b0be1aea1b8b3dcfece16cb57e7fc4aa407f83c59a275323149aac85e163e8b"]
-ERC777Transactions = ["0x4ddef98703d376e87db02337f94ee6bd0406ccfc11bf221ff8d9683513c53967", "0xb460976f3838d73067d949ba9a32635423d72e3b6a40cffb25e813ddee6ee1a8", "0xbe2bb6152f614f2a0258c9b16725039599f1740526a2c848e765ef0921c6a02d", "0x96bb6802d2d1da87555565b1e368c9042c6ebd9dd872a818ec876bdd8411d5a4", "0x6e08c5f80f527a1f7508c798d8eab08e103bc9591044149c70143b1a8f8270e7"]
+ERC20Transactions = ["0x45f2d5adf0e968b789d96dcec77ec097be84c30d73ba116fd8d0d361dfa0a597", "0x162182ce72764c021fe2042bdd037f4d0cfbaf248acd0b16e83a9bc60ade3623", "0xb9c2beac5baedf4b17a0c6f00df0f2b67fd24b423e1462b7cdda1fadc8147d4d", "0x2a9c8417dca16f7026fa465f7bb2cd6d7e845ea0dd773cfd961647381da0c5d8", "0x3b0be1aea1b8b3dcfece16cb57e7fc4aa407f83c59a275323149aac85e163e8b"]
+ERC777Transactions = ["0x4990024c40c01d05863cf272961c77ebfb07f0b7ab075937ca77d815db133c8a", "0x10a13c3b9a3275b8ad7f2ab419b99ad39f6e2a3f0bdefe55b4d5696141956b29", "0x78a5e1b7736b7e4a6d3fedd1110c420d7c8421a99cf3214e79e4ba8a9e9938c1", "0x96bb6802d2d1da87555565b1e368c9042c6ebd9dd872a818ec876bdd8411d5a4", "0x6e08c5f80f527a1f7508c798d8eab08e103bc9591044149c70143b1a8f8270e7"]
